requirements.txt
# Google API dependencies
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0

# Data processing
pandas==2.1.1

# Sentiment analysis
textblob==0.17.1

# Visualization - NEW
matplotlib==3.8.0
seaborn==0.13.0